vibration-analysis/requirements.txt
paho-mqtt
influxdb-client
numpy
scipy
pandas
scikit-learn
joblib